+++ mad/optimize/_goal_allocation.py
@@ -1,32 +1,12 @@
 from mad.data_structures import GoalNode
-from typing import Dict, List
+from typing import Dict
 
 # private function should be as follows
-def _get_goals(goal_tree: GoalNode) -> List:
-    """
-    Takes in a goal tree and traverses it using BFS and appends each GoalNode to an output list
-
-    Parameters
-    ----------
-    goal_tree : mad.data_structures.GoalNode
-        Hierarchical Multi Agent Goal Tree
-    """
-    output = []
-    q = []
-    q.append(goal_tree)
-    while q:
-        current = q[0]
-        q.pop(0)
-        output.append(current)
-        for child in current.get_children():
-            q.append(child)
-
-    return output
-
+def _helper_func():
+    pass
 
 def initial_goal_allocation(goal_tree: GoalNode,
-                            max_resources: int,
-                            agents: List) -> Dict:
+                            max_resources: int) -> Dict:
     
     """
     Optimizes allocation of goals to multiple agents
@@ -44,35 +24,58 @@ def initial_goal_allocation(goal_tree: GoalNode,
         Allocates list of goals (value) to each agent (key)
     """
 
-    # Write your code here
+    # write your code here
 
     # Raise an error if goal_tree is empty
 
+    pass
 
-def jonathan_average_cost(goal_tree: GoalNode) -> None:
+
+
+"""
+Jonathan's Algorithm
+########################################################
+"""
+from typing import Dict, List
+from mad.data_structures import GoalNode
+from mad.optimize import jonathan_average_cost
+from mad.optimize import jonathan_optimal_path
+from mad.optimize import jonathan_distribute_goals
+
+def jonathan_algorithm(goal_tree: GoalNode, agents: List, max_resources: int) -> Dict:
     """
-    Takes in a goal tree and updates each GoalNode's agent cost to a temporary value of the average cost of all agents able to accomplish that goal
+    Takes in a goal tree and finds optimal goals to accomplish the main goal and distributes them to agents evenly
 
     Parameters
     ----------
     goal_tree : mad.data_structures.GoalNode
         Hierarchical Multi Agent Goal Tree
+    agents : List
+        List of string names of agents available
+    max_resources : int
+        Value for the max amount of resources each agent has available
+
+    Returns
+    -------
+    Returns : Dict
+        Dictionary of agent names (keys) and list of GoalNodes assigned (values)
     """
+
+    # Takes in a goal tree and updates each GoalNode's agent cost to a temporary value of the average cost of all agents able to accomplish that goal
+    jonathan_average_cost(goal_tree)
+
+    # Takes in a goal tree and max resources for each agent and finds the most optimal goal path based on the GoalNode.cost values through out the tree and returns a list of GoalNodes that should be accomplished
+    selected_goals = jonathan_optimal_path(goal_tree, max_resources)
+    print("Selected Goals:")
+    for goal in selected_goals:
+        print(goal.name)
     
-    # Raise an error if goal_tree is empty (???)
-    if goal_tree is None:
-        raise ValueError("Tree is empty!")
+    print()
+    # Takes in a list of GoalNodes and distributes them among available agents
+    distributed_goals = jonathan_distribute_goals(selected_goals, agents, max_resources)
 
-    # For each goal find average cost among available agents and assign the temporary cost to the goal
-    goals = _get_goals(goal_tree)
+    return distributed_goals
 
-    for goal in goals:
-        costs = [x for x in goal.data.values()]
-        avg_cost = sum(costs) / len(costs)
-        goal.cost = avg_cost
-        
-
-
-
-
-
+"""
+########################################################
+"""
