+++ mad/optimize/jonathan_algo_components/_path_optimization.py
