--- mad/optimize/tests/test_goal_allocation.py
@@ -1,16 +1,0 @@
-import pytest
-
-from mad.data_structures import GoalNode
-from mad.optimize import initial_goal_allocation
-
-
-def test_tree_creation():
-
-    # write your test case here
-    pass 
-
-def test_goal_allocation():
-
-    # write your test case here
-    pass
-
